feat: Add Redux store and provider to main component
This commit adds the necessary code changes to integrate Redux into the main component of the application. The Redux store is imported from the newly created "store.ts" file, and the Provider component from the "react-redux" library is wrapped around the App component. This allows the entire application to have access to the Redux store and enables the usage of Redux features such as state management and actions.